Move `BIGNUM` code into new `bn` module.
I agree to license my contributions to each file under the terms given
at the top of each file I changed.
